backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
pytest
pytest-django
coverage
django-cors-headers